FRASENTINEL/001/FRA-SENTINEL/requirements_enhanced.txt
attrs==25.3.0
base58==2.1.1
blinker==1.9.0
certifi==2025.8.3
cffi==2.0.0
chardet==3.0.4
charset-normalizer==3.4.3
click==8.2.1
click-plugins==1.1.1.2
cligj==0.7.2
colorama==0.4.6
cryptography==41.0.7
fiona==1.10.1
Flask==2.3.3
Flask-Babel==4.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
geopandas==0.14.1
googletrans==4.0.0rc1
greenlet==3.2.4
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
ipfshttpclient==0.8.0a2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multiaddr==0.0.9
netaddr==1.3.0
numpy==2.3.3
opencv-python==4.8.1.78
packaging==25.0
pandas==2.3.2
pdf2image==1.16.3
pillow==11.3.0
plotly==5.17.0
pycparser==2.23
PyJWT==2.10.1
pypng==0.20220715.0
pyproj==3.7.2
pytesseract==0.3.10
python-dateutil==2.9.0.post0
pytz==2025.2
qrcode==7.4.2
requests==2.32.5
rfc3986==1.5.0
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
tenacity==9.1.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
varint==1.0.2
Werkzeug==3.1.3
spacy>=3.4.0
nltk>=3.8.0
scikit-image>=0.20.0
scikit-learn>=1.3.0
